Fix debug build: EncryptedSharedPreferences 1.0 API and FIT calories type.
Unblocks ./gradlew test assembleDebug after the first compile pass.

+++ app/src/main/java/com/watchrelay/core/importing/FitParser.kt
@@ -53,7 +53,7 @@ object FitParser {
                             5 -> sport = SportMapper.fromFit(value.toInt())
                             7 -> if (value > 0) durationMillis = value
                             9 -> distanceMeters = value / 100.0
-                            11 -> calories = value
+                            11 -> calories = value.toDouble()
                         }
                     }
                 }
